Fix typo in README.md title
The typo in the project title within the README.md file has been fixed. It was changing the meaning and confusing readers about the project's actual role. It has now been modified from "Omni Drivenn (Infrastructure)" to "Omni Driven (Infrastructure)" to accurately reflect its role as the infrastructure module in the Omni project.